+++ components/table/Table.vue
@@ -4,37 +4,27 @@ import Empty from '../empty'
 import Pagination from '../pagination'
 interface Column {
   title?: string // 列头显示文字
-  width: number | string // 列宽度
+  width?: number | string // 列宽度，单位 px
   dataIndex: string // 列数据字符索引
   slot?: string // 列插槽名称索引
-}
-interface Pagination {
-  // 具体可参考 Pagination 分页组件相关 APIs
-  page?: number // 当前页码
-  pageSize?: number // 每页条数
-  pageSizeOptions?: string[] | number[] // 每页可以显示多少条
-  pageListNum?: number // 显示的页码数组长度
-  hideOnSinglePage?: boolean // 只有一页时是否隐藏分页
-  showQuickJumper?: boolean // 是否可以快速跳转至某页
-  showSizeChanger?: boolean // 是否展示 pageSize 切换器，当 total 大于 50 时默认为 true
-  showTotal?: boolean // 是否显示当前页数和数据总量
-  placement?: 'left' | 'center' | 'right' // 分页展示位置：靠左、居中、靠右
 }
 interface Props {
   columns?: Column[] // 表格列的配置项
   dataSource?: any[] // 表格数据数组
-  pagination?: Pagination // 分页配置
-  showPagination?: boolean // 是否显示分页
-  total?: number // 数据总数
   loading?: boolean // 是否加载中
+  spinProps?: object // Spin 组件属性配置，参考 Spin Props
+  emptyProps?: object // Empty 组件属性配置，参考 Empty Props
+  showPagination?: boolean // 是否显示分页
+  pagination?: object // Pagination 组件属性配置，参考 Pagination Props
 }
 withDefaults(defineProps<Props>(), {
   columns: () => [],
   dataSource: () => [],
-  pagination: () => ({}),
+  loading: false,
+  spinProps: () => ({}),
+  emptyProps: () => ({}),
   showPagination: true,
-  total: 0,
-  loading: false
+  pagination: () => ({})
 })
 const emit = defineEmits(['change'])
 function changePage(page: number, pageSize: number) {
@@ -59,11 +49,11 @@ function changePage(page: number, pageSize: number) {
       </thead>
       <tbody class="m-body">
         <tr class="m-tr-loading" v-show="loading">
-          <Spin class="m-loading" size="small" :colspan="columns.length" />
+          <Spin class="m-loading" size="small" v-bind="spinProps" :colspan="columns.length" />
         </tr>
-        <tr class="m-tr-empty" v-show="!total">
+        <tr class="m-tr-empty" v-show="!dataSource.length">
           <td class="m-td-empty" :colspan="columns.length">
-            <Empty class="empty" image="outlined" />
+            <Empty class="empty" image="outlined" v-bind="emptyProps" />
           </td>
         </tr>
         <tr class="m-tr" v-for="(data, index) in dataSource" :key="index">
@@ -76,21 +66,7 @@ function changePage(page: number, pageSize: number) {
         </tr>
       </tbody>
     </table>
-    <Pagination
-      class="mt20"
-      @change="changePage"
-      :total="total"
-      :page="pagination.page"
-      :pageSize="pagination.pageSize"
-      :pageSizeOptions="pagination.pageSizeOptions"
-      :pageListNum="pagination.pageListNum"
-      :hideOnSinglePage="pagination.hideOnSinglePage"
-      :showQuickJumper="pagination.showQuickJumper"
-      :showSizeChanger="pagination.showSizeChanger"
-      :showTotal="pagination.showTotal"
-      :placement="pagination.placement"
-      v-if="showPagination && total"
-    />
+    <Pagination class="mt20" v-bind="pagination" @change="changePage" v-if="showPagination" />
   </div>
 </template>
 <style lang="less" scoped>